# [Deepin-Kernel-SIG] [linux 6.6-y] [Upstream] LoongArch: Relax memory ordering for atomic operations