@opsiff 2025-04-07
Before:
8 CPUs in system; running 1 parallel copy of tests
Dhrystone 2 using register variables       41193038.5 lps   (10.0 s, 7 samples)
Double-Precision Whetstone                     5035.1 MWIPS (10.0 s, 7 samples)
Execl Throughput                               4248.3 lps   (30.0 s, 2 samples)
File Copy 1024 bufsize 2000 maxblocks        621670.0 KBps  (30.0 s, 2 samples)
File Copy 256 bufsize 500 maxblocks          164863.5 KBps  (30.0 s, 2 samples)
File Copy 4096 bufsize 8000 maxblocks       1896980.0 KBps  (30.0 s, 2 samples)
Pipe Throughput                             1187258.1 lps   (10.0 s, 7 samples)
Pipe-based Context Switching                 109270.7 lps   (10.0 s, 7 samples)
Process Creation                               8702.0 lps   (30.0 s, 2 samples)
Shell Scripts (1 concurrent)                  11797.6 lpm   (60.0 s, 2 samples)
Shell Scripts (8 concurrent)                   4773.6 lpm   (60.0 s, 2 samples)
System Call Overhead                        1284812.1 lps   (10.0 s, 7 samples)
System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0   41193038.5   3529.8
Double-Precision Whetstone                       55.0       5035.1    915.5
Execl Throughput                                 43.0       4248.3    988.0
File Copy 1024 bufsize 2000 maxblocks          3960.0     621670.0   1569.9
File Copy 256 bufsize 500 maxblocks            1655.0     164863.5    996.2
File Copy 4096 bufsize 8000 maxblocks          5800.0    1896980.0   3270.7
Pipe Throughput                               12440.0    1187258.1    954.4
Pipe-based Context Switching                   4000.0     109270.7    273.2
Process Creation                                126.0       8702.0    690.6
Shell Scripts (1 concurrent)                     42.4      11797.6   2782.5
Shell Scripts (8 concurrent)                      6.0       4773.6   7956.1
System Call Overhead                          15000.0    1284812.1    856.5
System Benchmarks Index Score                                        1398.0
Dhrystone 2 using register variables      221318057.8 lps   (10.0 s, 7 samples)
Double-Precision Whetstone                    37174.1 MWIPS (10.0 s, 7 samples)
Execl Throughput                              23089.8 lps   (30.0 s, 2 samples)
File Copy 1024 bufsize 2000 maxblocks       3426413.9 KBps  (30.0 s, 2 samples)
File Copy 256 bufsize 500 maxblocks          936160.4 KBps  (30.0 s, 2 samples)
File Copy 4096 bufsize 8000 maxblocks       7895970.5 KBps  (30.0 s, 2 samples)
Pipe Throughput                             6994333.2 lps   (10.0 s, 7 samples)
Pipe-based Context Switching                 967203.7 lps   (10.0 s, 7 samples)
Process Creation                              48383.1 lps   (30.0 s, 2 samples)
Shell Scripts (1 concurrent)                  41782.5 lpm   (60.0 s, 2 samples)
Shell Scripts (8 concurrent)                   5426.0 lpm   (60.0 s, 2 samples)
System Call Overhead                        7489205.5 lps   (10.0 s, 7 samples)
System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0  221318057.8  18964.7
Double-Precision Whetstone                       55.0      37174.1   6758.9
Execl Throughput                                 43.0      23089.8   5369.7
File Copy 1024 bufsize 2000 maxblocks          3960.0    3426413.9   8652.6
File Copy 256 bufsize 500 maxblocks            1655.0     936160.4   5656.6
File Copy 4096 bufsize 8000 maxblocks          5800.0    7895970.5  13613.7
Pipe Throughput                               12440.0    6994333.2   5622.5
Pipe-based Context Switching                   4000.0     967203.7   2418.0
Process Creation                                126.0      48383.1   3839.9
Shell Scripts (1 concurrent)                     42.4      41782.5   9854.4
Shell Scripts (8 concurrent)                      6.0       5426.0   9043.3
System Call Overhead                          15000.0    7489205.5   4992.8
System Benchmarks Index Score                                        6848.9
After:
8 CPUs in system; running 1 parallel copy of tests
Dhrystone 2 using register variables       41338050.2 lps   (10.0 s, 7 samples)
Double-Precision Whetstone                     5035.5 MWIPS (10.0 s, 7 samples)
Execl Throughput                               4290.3 lps   (30.0 s, 2 samples)
File Copy 1024 bufsize 2000 maxblocks        625513.1 KBps  (30.0 s, 2 samples)
File Copy 256 bufsize 500 maxblocks          166766.0 KBps  (30.0 s, 2 samples)
File Copy 4096 bufsize 8000 maxblocks       1892782.1 KBps  (30.0 s, 2 samples)
Pipe Throughput                             1188553.2 lps   (10.0 s, 7 samples)
Pipe-based Context Switching                 110585.7 lps   (10.0 s, 7 samples)
Process Creation                               8580.5 lps   (30.0 s, 2 samples)
Shell Scripts (1 concurrent)                  11865.7 lpm   (60.0 s, 2 samples)
Shell Scripts (8 concurrent)                   4810.9 lpm   (60.0 s, 2 samples)
System Call Overhead                        1285163.6 lps   (10.0 s, 7 samples)
System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0   41338050.2   3542.2
Double-Precision Whetstone                       55.0       5035.5    915.5
Execl Throughput                                 43.0       4290.3    997.7
File Copy 1024 bufsize 2000 maxblocks          3960.0     625513.1   1579.6
File Copy 256 bufsize 500 maxblocks            1655.0     166766.0   1007.6
File Copy 4096 bufsize 8000 maxblocks          5800.0    1892782.1   3263.4
Pipe Throughput                               12440.0    1188553.2    955.4
Pipe-based Context Switching                   4000.0     110585.7    276.5
Process Creation                                126.0       8580.5    681.0
Shell Scripts (1 concurrent)                     42.4      11865.7   2798.5
Shell Scripts (8 concurrent)                      6.0       4810.9   8018.1
System Call Overhead                          15000.0    1285163.6    856.8
System Benchmarks Index Score                                        1402.9
Dhrystone 2 using register variables      221329300.0 lps   (10.0 s, 7 samples)
Double-Precision Whetstone                    37173.5 MWIPS (10.0 s, 7 samples)
Execl Throughput                              23274.5 lps   (30.0 s, 2 samples)
File Copy 1024 bufsize 2000 maxblocks       3462085.0 KBps  (30.0 s, 2 samples)
File Copy 256 bufsize 500 maxblocks          954300.1 KBps  (30.0 s, 2 samples)
File Copy 4096 bufsize 8000 maxblocks       7714182.9 KBps  (30.0 s, 2 samples)
Pipe Throughput                             7003171.4 lps   (10.0 s, 7 samples)
Pipe-based Context Switching                 994933.1 lps   (10.0 s, 7 samples)
Process Creation                              48563.7 lps   (30.0 s, 2 samples)
Shell Scripts (1 concurrent)                  41691.9 lpm   (60.0 s, 2 samples)
Shell Scripts (8 concurrent)                   5477.6 lpm   (60.0 s, 2 samples)
System Call Overhead                        7492018.1 lps   (10.0 s, 7 samples)

System Benchmarks Index Values               BASELINE       RESULT    INDEX
Dhrystone 2 using register variables         116700.0  221329300.0  18965.7
Double-Precision Whetstone                       55.0      37173.5   6758.8
Execl Throughput                                 43.0      23274.5   5412.7
File Copy 1024 bufsize 2000 maxblocks          3960.0    3462085.0   8742.6
File Copy 256 bufsize 500 maxblocks            1655.0     954300.1   5766.2
File Copy 4096 bufsize 8000 maxblocks          5800.0    7714182.9  13300.3
Pipe Throughput                               12440.0    7003171.4   5629.6
Pipe-based Context Switching                   4000.0     994933.1   2487.3
Process Creation                                126.0      48563.7   3854.3
Shell Scripts (1 concurrent)                     42.4      41691.9   9833.0
Shell Scripts (8 concurrent)                      6.0       5477.6   9129.4
System Call Overhead                          15000.0    7492018.1   4994.7
System Benchmarks Index Score                                        6880.5